see #2107: Updated fwRuntime: remove warning (WARNINGSASERRORS mode)
--HG--
branch : fw4spl_0.8.3

+++ SrcLib/core/fwRuntime/src/ConfigurationElementContainer.cpp
@@ -106,7 +106,7 @@ ConfigurationElementContainer ConfigurationElementContainer::findAllConfiguratio
 
 //-----------------------------------------------------------------------------
 
-unsigned int ConfigurationElementContainer::size() const
+size_t ConfigurationElementContainer::size() const
 {
     return m_elements.size();
 }
